Cleanup router for which processing added router failed
In the _process_added_router() method of the L3 agent, if processing
router will fail, router_info should be cleaned to e.g. be removed from
the router cache so it will not be treated as updated router in next
iteration of the agent.

Conflicts:
    neutron/agent/l3/agent.py

Closes-Bug: #1947993
(cherry picked from commit 41159bd9a4178a5f8a4fdc90b034c48fa9880cac)
(cherry picked from commit 1c17019f6c7957c95b1d12942c5898fcc111e4ab)

+++ neutron/agent/l3/agent.py
@@ -484,17 +484,21 @@ class L3NATAgent(ha.AgentMixin,
             ri.initialize(self.process_monitor)
         except Exception:
             with excutils.save_and_reraise_exception():
-                del self.router_info[router_id]
                 LOG.exception('Error while initializing router %s',
                               router_id)
-                self.namespaces_manager.ensure_router_cleanup(router_id)
-                try:
-                    ri.delete()
-                except Exception:
-                    LOG.exception('Error while deleting router %s',
-                                  router_id)
+                self._cleanup_failed_router(router_id, delete_router_info=True)
 
         self._resize_process_pool()
+
+    def _cleanup_failed_router(self, router_id, delete_router_info):
+        ri = self.router_info.pop(router_id)
+        self.namespaces_manager.ensure_router_cleanup(router_id)
+        try:
+            if delete_router_info:
+                ri.delete()
+        except Exception:
+            LOG.exception('Error while deleting router %s',
+                          router_id)
 
     def _safe_router_removed(self, router_id):
         """Try to delete a router and return True if successful."""
@@ -624,8 +628,22 @@ class L3NATAgent(ha.AgentMixin,
         self._router_added(router['id'], router)
         ri = self.router_info[router['id']]
         ri.router = router
-        ri.process()
+        try:
+            ri.process()
+        except Exception:
+            with excutils.save_and_reraise_exception():
+                LOG.exception('Error while processing router %s',
+                              router['id'])
+                # NOTE(slaweq): deleting of the router info in the
+                # _cleanup_failed_router is avoided as in case of error,
+                # processing of the router will be retried on next call and
+                # that may lead to some race conditions e.g. with
+                # configuration of the DVR router's FIP gateway
+                self._cleanup_failed_router(router['id'],
+                                            delete_router_info=False)
+
         registry.notify(resources.ROUTER, events.AFTER_CREATE, self, router=ri)
+
         self.l3_ext_manager.add_router(self.context, router)
 
     def _process_updated_router(self, router):
